Use SvRXOK() on 5.12+ instead of SVt_REGEXP in XS version:
Even though SVt_REGEXP exists, SvRXOK() is better because it
adds a checking for magic. While we technically don't care about
magic, regular expressions are different and this has to be taken
into account.

This is the first step in fixing the regexp part.

The next parts will involve updating the opcode version to use
this as well (which will be a special case) and to support SvRXOK
in older versions of perl (5.8 and below).