Behavior change: is_scalarref($refref) is now false
The rationale for this is two-fold. First, callers that want to decide how
to handle inputs based on their reference type will usually want to treat a
ref-ref and a scalar-ref differently. Secondly, this more closely matches
the behavior of the ref builtin and of Scalar::Util::reftype(), which both
report a ref-ref as "REF" rather than "SCALAR".