Behavior change: is_scalarref() false for qr/./ and \*GLOB
The rationale for this is essentially the same as for ref-ref. First, callers
that want to decide how to handle inputs based on their reference type will
usually want to treat a regexp or a glob-ref differently from a scalar-ref.
Secondly, this more closely matches the behavior of the ref builtin (which
reports these cases as "Regexp" and "GLOB" respectively, rather than "SCALAR")
and of Scalar::Util::reftype() (likewise, but "REGEXP" and "GLOB").